backend/requirements.txt
Flask==3.0.2
flask-cors==4.0.0
google-generativeai==0.4.1
python-dotenv==1.0.1
Flask-Mail==0.9.1
